% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{add_h5_list_adt_results}
\alias{add_h5_list_adt_results}
\title{Add ADT results to an h5_list.}
\usage{
add_h5_list_adt_results(
  h5_list,
  adt_metadata = NULL,
  adt_matrix,
  match_target = "barcodes"
)
}
\arguments{
\item{h5_list}{a list object generated by running rhdf5::h5dump() on a 10x HDF5 file.}

\item{adt_metadata}{(optional) a data.frame or data.table containing ADT metadata. The first column should contain cell identifiers (i.e. barcodes) that match the rownames of the adt_matrix.}

\item{adt_matrix}{a matrix or dgCMatrix containing ADT counts. Rows should be cells, columns should be features. rownames should be cell IDs (i.e. barcodes)}

\item{match_target}{a character object specifying which values to use for filtering in the h5_list. If "barcodes", uses matrix/barcodes; Otherwise, looks in matrix/observations for matching values}
}
\value{
a list object with ADT data added to /ADT/ and /matrix/observations/
}
\description{
Add ADT results to an h5_list.
}
