% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_h5ad.R
\name{read_h5ad_cell_meta}
\alias{read_h5ad_cell_meta}
\title{Read .h5ad Cell Metadata}
\usage{
read_h5ad_cell_meta(h5ad_file)
}
\arguments{
\item{h5ad_file}{the path to an .h5ad file}
}
\value{
A data.frame containing all feature metadata found in /obs
}
\description{
Read .h5ad Cell Metadata
}
