% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_h5ad.R
\name{read_h5ad_dgCMatrix}
\alias{read_h5ad_dgCMatrix}
\title{Read the from a .h5ad file as a sparse matrix}
\usage{
read_h5ad_dgCMatrix(h5ad_file, feature_names = "name")
}
\arguments{
\item{h5ad_file}{the path to an .h5ad file}

\item{feature_names}{a character object specifying whether to use "id" or "name" for row.names. Default is "name".}
}
\value{
a sparse matrix of class Matrix::dgCMatrix

a dgCMatrix of gene expression values.
}
\description{
Read the from a .h5ad file as a sparse matrix
}
