% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{read_h5ad_sample_meta}
\alias{read_h5ad_sample_meta}
\title{Read .h5 Cell SampleInfo}
\usage{
read_h5ad_sample_meta(h5ad_file)
}
\arguments{
\item{h5_file}{the path to an .h5 file in 10x Genomics format}

\item{target}{A matrix object in the .h5 file with a /sample object and/or a /target/observations/ sub-group. Default is "matrix".}
}
\value{
A data.frame containing all sample metadata found in /target/sample and /target/observations/
}
\description{
Read .h5 Cell SampleInfo
}
